add Interface for multiple parameter store files. store_aws for AWS, store_config for local config files.
